add moviesItemDecoration
